third_party/WebKit/LayoutTests/fast/html/imports/import-onerror.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
<script>
testRunner.waitUntilDone();

var givenLoadEventTargets = [];
function handleError(event)
{
    givenLoadEventTargets.push(event.target.id);
    if (givenLoadEventTargets.length == 3) {
        shouldBe("['invalid', 'notFound', 'notFoundShared']", "givenLoadEventTargets.sort()");
        testRunner.notifyDone();
    }
}
</script>
<link id="notFound" rel="import" href="notfound.html" onerror="handleError(event)">
<link id="notFoundShared" rel="import" href="notfound.html" onerror="handleError(event)">
<link id="invalid" rel="import" href="" onerror="handleError(event)">
</head>
<body></body>
<script src="../../js/resources/js-test-post.js"></script>
</html>
